+++ es/provider.go
@@ -5,6 +5,7 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
+	"fmt"
 	"log"
 	"net/http"
 	"net/url"
@@ -35,6 +36,8 @@ type ProviderConf struct {
 	cacertFile         string
 	username           string
 	password           string
+	token              string
+	tokenName          string
 	parsedUrl          *url.URL
 	signAWSRequests    bool
 	esVersion          string
@@ -80,6 +83,18 @@ func Provider() terraform.ResourceProvider {
 				Optional:    true,
 				DefaultFunc: schema.EnvDefaultFunc("ELASTICSEARCH_PASSWORD", nil),
 				Description: "Password to use to connect to elasticsearch using basic auth",
+			},
+			"token": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				DefaultFunc: schema.EnvDefaultFunc("ELASTICSEARCH_TOKEN", nil),
+				Description: "A bearer token or ApiKey for an Authorization header, e.g. Active Directory API key.",
+			},
+			"token_name": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     "ApiKey",
+				Description: "The type of token, usually ApiKey or Bearer",
 			},
 			"aws_assume_role_arn": {
 				Type:        schema.TypeString,
@@ -216,6 +231,8 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 		cacertFile:      d.Get("cacert_file").(string),
 		username:        d.Get("username").(string),
 		password:        d.Get("password").(string),
+		token:           d.Get("token").(string),
+		tokenName:       d.Get("token_name").(string),
 		parsedUrl:       parsedUrl,
 		signAWSRequests: d.Get("sign_aws_requests").(bool),
 		esVersion:       d.Get("elasticsearch_version").(string),
@@ -254,6 +271,8 @@ func getClient(conf *ProviderConf) (interface{}, error) {
 		opts = append(opts, elastic7.SetHttpClient(awsHttpClient(awsRegion, conf)), elastic7.SetSniff(false))
 	} else if conf.insecure || conf.cacertFile != "" {
 		opts = append(opts, elastic7.SetHttpClient(tlsHttpClient(conf)), elastic7.SetSniff(false))
+	} else if conf.token != "" {
+		opts = append(opts, elastic7.SetHttpClient(tokenHttpClient(conf.token, conf.tokenName, conf.insecure)), elastic7.SetSniff(false))
 	}
 
 	var relevantClient interface{}
@@ -298,7 +317,10 @@ func getClient(conf *ProviderConf) (interface{}, error) {
 			opts = append(opts, elastic6.SetHttpClient(awsHttpClient(awsRegion, conf)), elastic6.SetSniff(false))
 		} else if conf.insecure || conf.cacertFile != "" {
 			opts = append(opts, elastic6.SetHttpClient(tlsHttpClient(conf)), elastic6.SetSniff(false))
+		} else if conf.token != "" {
+			opts = append(opts, elastic6.SetHttpClient(tokenHttpClient(conf.token, conf.tokenName, conf.insecure)), elastic6.SetSniff(false))
 		}
+
 		relevantClient, err = elastic6.NewClient(opts...)
 		if err != nil {
 			return nil, err
@@ -327,7 +349,10 @@ func getClient(conf *ProviderConf) (interface{}, error) {
 			opts = append(opts, elastic5.SetHttpClient(awsHttpClient(awsRegion, conf)), elastic5.SetSniff(false))
 		} else if conf.insecure || conf.cacertFile != "" {
 			opts = append(opts, elastic5.SetHttpClient(tlsHttpClient(conf)), elastic5.SetSniff(false))
+		} else if conf.token != "" {
+			opts = append(opts, elastic5.SetHttpClient(tokenHttpClient(conf.token, conf.tokenName, conf.insecure)), elastic5.SetSniff(false))
 		}
+
 		relevantClient, err = elastic5.NewClient(opts...)
 		if err != nil {
 			return nil, err
@@ -339,9 +364,12 @@ func getClient(conf *ProviderConf) (interface{}, error) {
 	return relevantClient, nil
 }
 
-func assumeRoleCredentials(region, roleARN string) *awscredentials.Credentials {
-	sess := awssession.Must(awssession.NewSession(&aws.Config{
-		Region: aws.String(region),
+func assumeRoleCredentials(region, roleARN, profile string) *awscredentials.Credentials {
+	sess := awssession.Must(awssession.NewSessionWithOptions(awssession.Options{
+		Profile: profile,
+		Config: aws.Config{
+			Region: aws.String(region),
+		},
 	}))
 	stsClient := awssts.New(sess)
 	assumeRoleProvider := &awsstscreds.AssumeRoleProvider{
@@ -367,7 +395,7 @@ func awsSession(region string, conf *ProviderConf) *awssession.Session {
 	if conf.awsAccessKeyId != "" {
 		sessOpts.Config.Credentials = awscredentials.NewStaticCredentials(conf.awsAccessKeyId, conf.awsSecretAccessKey, conf.awsSessionToken)
 	} else if conf.awsAssumeRoleArn != "" {
-		sessOpts.Config.Credentials = assumeRoleCredentials(region, conf.awsAssumeRoleArn)
+		sessOpts.Config.Credentials = assumeRoleCredentials(region, conf.awsAssumeRoleArn, conf.awsProfile)
 	} else if conf.awsProfile != "" {
 		sessOpts.Profile = conf.awsProfile
 	}
@@ -385,13 +413,29 @@ func awsSession(region string, conf *ProviderConf) *awssession.Session {
 
 func awsHttpClient(region string, conf *ProviderConf) *http.Client {
 	signer := awssigv4.NewSigner(awsSession(region, conf).Config.Credentials)
-	client, _ := aws_signing_client.New(signer, nil, "es", region)
+	client, err := aws_signing_client.New(signer, nil, "es", region)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	return client
+}
+
+func tokenHttpClient(token string, tokenName string, insecure bool) *http.Client {
+	client := http.DefaultClient
+
+	rt := WithHeader(client.Transport)
+	rt.Set("Authorization", fmt.Sprintf("%s %s", tokenName, token))
+	client.Transport = rt
+
+	if insecure {
+		client.Transport.(*http.Transport).TLSClientConfig.InsecureSkipVerify = true
+	}
 
 	return client
 }
 
 func tlsHttpClient(conf *ProviderConf) *http.Client {
-
 	// Configure TLS/SSL
 	tlsConfig := &tls.Config{}
 	if conf.certPemPath != "" && conf.keyPemPath != "" {